pyproject.toml
[tool.poetry]
name = "capacities-companion"
version = "0.1.0"
description = "A suite of companion tools for Capacities (the PKM app)"
authors = ["Trevor Hubbard <trevormhubbard@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
networkx = "^3.5"
pyyaml = "^6.0.2"
tqdm = "^4.67.1"
openai = "^1.99.0"
ollama = "^0.5.1"
pydantic = "^2.11.7"
rich = "^14.1.0"
python-dotenv = "^1.1.1"
lxml = "^6.0.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pickleshare = "^0.7.5"
nbformat = "^5.10.4"
ipycytoscape = "^1.3.3"

